Restructured main module code to be more readable Scoped extension CSS variables to prevent conflicts
