+++ src/popup/popup.html
@@ -1,8 +1,8 @@
-<!DOCTYPE html>
-<html>
+<!doctype html>
+<html class="tz-initialized">
   <head>
-    <meta charset="UTF-8">
-    <link rel="stylesheet" href="popup.css">
+    <meta charset="UTF-8" />
+    <link rel="stylesheet" href="popup.css" />
     <script src="popup.ts" type="module" defer></script>
   </head>
   <body>
