+++ src/styles/variables.css
@@ -1,4 +1,4 @@
-:root {
+.tz-initialized {
   --color-primary: #82c7ff;
   --color-secondary: orange;
   --color-tertiary: blueviolet;
